Fixed Alsa and AO sounds, and FIRFilter member initialization order

+++ Source/Core/AudioCommon/AlsaSoundStream.h
@@ -54,7 +54,7 @@ private:
 	bool AlsaInit();
 	void AlsaShutdown();
 
-	s16 mix_buffer[BUFFER_SIZE_MAX * CHANNEL_COUNT];
+	float mix_buffer[BUFFER_SIZE_MAX * CHANNEL_COUNT];
 	std::thread thread;
 	std::atomic<ALSAThreadStatus> m_thread_status;
 	std::condition_variable cv;
